+++ life/runtime/runtime_test.go
@@ -20,9 +20,6 @@ func TestDefaults(t *testing.T) {
 	cfg := new(Config)
 	setDefaults(cfg)
 
-	if cfg.Difficulty == nil {
-		t.Error("expected difficulty to be non nil")
-	}
 	if cfg.Time == nil {
 		t.Error("expected time to non nil")
 	}